implementando endpoint de criação de material

+++ Martins.Backend.Domain/Models/Material.cs
@@ -9,11 +9,11 @@ namespace Martins.Backend.Domain.Models
         public Guid Id { get; set; }
         public required string Name { get; set; }
         public required string Category { get; set; }
-        public int CurrentStock { get; set; }
+        public decimal CurrentStock { get; set; }
         public UnitOfMeasureEnum UnitOfMeasure { get; set; }
-        public int TotalCost { get; set; }
-        public int UnitCost { get; set; }
-        public int LowStockThreshold { get; set; }
+        public decimal TotalCost { get; set; }
+        public decimal UnitCost { get; set; }
+        public decimal LowStockThreshold { get; set; }
         public DateTime LastUpdatedAt { get; set; }
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
         public Guid? SupplierId { get; set; }
